+++ src/js/modules/toast.js
@@ -12,6 +12,7 @@ export function create(config = {}) {
 
     const defaultConfig = {
         containerId: "tc-toast-container", // ID of toast container (parent element)
+        removeExisting: false, // Remove any existing toasts which exist inside the container
         toastId: "", // ID of toast (blank for none)
         classes: "tc-style-default", // Class(es) to add in addition to "tc-toast"
         duration: 250, // Animation duration in milliseconds
@@ -83,6 +84,10 @@ export function create(config = {}) {
 
     if (config.toastId !== "") {
         toast.id = config.toastId;
+    }
+
+    if (config.removeExisting) {
+        container.innerHTML = "";
     }
 
     container.appendChild(toast);