+++ CHANGELOG.md
@@ -13,6 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [2.0.0] - 2024.05.30
+
+### Added
+
+- Added `removeExisting` property on the toast config object to remove any existing toasts in the container
+- Added web components for icons, alerts, badges and toasts
+
+### Changed
+
+- Updated documentation
+- Updated dependencies
+
 ## [1.2.1] - 2023.08.02
 
 ### Fixed